Restyle air filling view and making it responsive
