+++ src/components/BlenderLogbook/components/SavingTile.tsx
@@ -1,9 +1,6 @@
 import { Field } from 'formik';
 import { ButtonType, PrimaryButton } from '../../common/Button/Buttons';
-import {
-  CommonTileProps,
-  LogbookCommonTileProps,
-} from '../NewBlenderFillingEvent';
+import { CommonTileProps } from '../../BlenderLogbook/NewBlenderFillingEvent';
 import React from 'react';
 
 type SavingTileProps = CommonTileProps & {
@@ -34,28 +31,6 @@ export const SavingTile: React.FC<SavingTileProps> = ({
       <PrimaryButton
         className="saveFillEventButton"
         disabled={!values.userConfirm || isSubmitting}
-        type={ButtonType.submit}
-        text="Tallenna täyttö"
-      />
-    </div>
-  </div>
-);
-
-export const AirLogbookSavingTile: React.FC<LogbookCommonTileProps> = ({
-  values,
-}): JSX.Element => (
-  <div className="tileWrapper savingTile">
-    <h2>Tallenna</h2>
-    <div>
-      <div className="savingFlexRow">
-        <Field id="id-userConfirm" type="checkbox" name="userConfirm" />
-        <label htmlFor="id-userConfirm">
-          Olen tarkistanut täyttämäni arvot
-        </label>
-      </div>
-      <PrimaryButton
-        className="saveFillEventButton"
-        disabled={!values.userConfirm}
         type={ButtonType.submit}
         text="Tallenna täyttö"
       />
