+++ src/components/Logbook/components/FillingTile.module.scss
@@ -1,0 +1,37 @@
+@import '../../../variables.scss';
+
+.cylinderSet {
+  background: $white-smoke-light;
+  border-radius: 12px;
+  border: 1px solid $timberwolf;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  gap: 1rem;
+  margin-bottom: 1rem;
+  padding: 0.5rem;
+  position: relative;
+}
+
+.deleteButtonWrapper {
+  align-items: flex-end;
+  display: flex;
+  position: absolute;
+  top: 0;
+  right: 0;
+
+  button {
+    background: transparent;
+    color: $red;
+
+    &:disabled {
+      color: $grey60;
+    }
+  }
+}
+
+.addRow {
+  button {
+    width: 100%;
+  }
+}
