+++ src/components/Logbook/components/SavingTile.module.scss
@@ -1,0 +1,5 @@
+.submit {
+  button {
+    width: 100%;
+  }
+}
