--- src/styles/blenderLogbook/blenderLogBook.css
@@ -1,68 +1,0 @@
-.tileWrapper {
-  background-color: #FFFFFF;
-  border-radius: 8px;
-  box-shadow: 0px 4px 10px 2px rgba(0, 0, 0, 0.1);
-  display: inline-block;
-  padding: 1rem;
-  position: relative;
-  width: 100%;
-}
-
-.fillingEventFlexRow {
-  display: flex;
-  flex-direction: row;
-  column-gap: 2rem;
-  padding-bottom: 2rem;
-}
-
-.savingFlexRow {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.pricingTile {
-  width: 320px;
-}
-
-.savingTile {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 520px;
-}
-
-.pricingGridRow {
-  display: grid;
-  grid-template-columns: auto 25% 20%;
-  margin: 0 0 0.25rem 0;
-  border-bottom: 1px solid #c3c3c3;
-}
-
-.basicInfoGridRow {
-  display: grid;
-  gap: 16px;
-  grid-template-columns: 220px 100px auto auto;
-  justify-content: space-between;
-  margin: 0 0 1rem 0;
-}
-
-.fillingEventGridRow {
-  align-items: flex-end;
-  display: grid;
-  grid-template-columns: 220px 220px 220px 100px 100px 64px;
-  justify-content: space-between;
-  margin: 0 0 1rem 0;
-}
-
-.addNewRow {
-  width: 100%;
-}
-
-.deleteRowButton {
-  height: 38px;
-}
-
-.saveFillEventButton {
-  width: 100%;
-}
