+++ src/variables.scss
@@ -5,6 +5,7 @@ $dark-blue: #003143;
 $blue: #526578;
 $blue-hover: #405c7a;
 $grey: #c3c3c3;
+$grey60: #999999;
 $white: #ffffff;
 $white-smoke: #f4f4f4;
 $white-smoke-light: #f4f4f6a6;
