+++ src/views/FillEvents.tsx
@@ -1,4 +1,3 @@
 import { ListFillEvents } from '../components/FillEvents/ListFillEvents';
-import '../styles/blenderLogbook/blenderLogBook.css';
 
 export const FillEvents = (): JSX.Element => <ListFillEvents />;
